@@ -22,7 +22,7 @@
     "build:es5:min": "uglifyjs dist/es5/uri.all.js --support-ie8 --output dist/es5/uri.all.min.js --in-source-map dist/es5/uri.all.js.map --source-map uri.all.min.js.map --comments --compress --mangle --pure-funcs merge subexp && mv uri.all.min.js.map dist/es5/ && cp dist/es5/uri.all.d.ts dist/es5/uri.all.min.d.ts",
     "build": "npm run build:esnext && npm run build:es5 && npm run build:es5:min",
     "clean": "rm -rf dist",
-    "test": "mocha -u mocha-qunit-ui dist/es5/uri.all.js tests/tests.js"
+    "test": "vitest --coverage"
   },
   "repository": {
     "type": "git",
@@ -62,14 +62,14 @@
     "babel-cli": "^6.26.0",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-preset-latest": "^6.24.1",
-    "mocha": "^8.2.1",
-    "mocha-qunit-ui": "^0.1.3",
+    "@vitest/coverage-v8": "^1.2.2",
     "rollup": "^0.41.6",
     "rollup-plugin-babel": "^2.7.1",
     "rollup-plugin-node-resolve": "^2.0.0",
     "sorcery": "^0.10.0",
-    "typescript": "^2.8.1",
-    "uglify-js": "^2.8.14"
+    "typescript": "^5.3.3",
+    "uglify-js": "^2.8.14",
+    "vitest": "^1.2.2"
   },
   "dependencies": {
     "punycode": "^2.1.0"
